pyproject.toml
[project]
authors = [
  {name = "redphx"},
  {name = "fourvrd"},
]
dependencies = [
  "construct==2.10.70",
  "hidapi==0.14.0.post4",
  "deepdiff==8.2.0",
  "python-dotenv==1.0.1",
]
description = "Python library to control Ulanzi D200 and D200H decks."
name = "UD200"
readme = "README.md"
requires-python = ">=3.9"
version = "0.1.0"

[tool.poetry]
packages = [{include = "ud200", from = "src"}]

[tool.poetry.group.dev.dependencies]
pytest = "==8.3.4"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=2.0.0"]
